Disable addition of simple-framebuffer by U-boot
In order to enable HDMI support for the Raspberry Pi, U-boot adds a
simple-framebuffer node to the device tree. However, this then prevents
the kernel from being able to correctly initialise the HDMI (which results
in a corrupted terminal), which subsequently prevents the camera from working.
Disabling HDMI support in U-boot resolves this issue. The U-boot console is
still available over the serial port.

Signed-off-by: Amr Bekhit <amrbekhit@gmail.com>